Conflicts:
	src/dr/util/Transform.java

+++ src/dr/inference/operators/AlternatingOperatorSchedule.java
@@ -124,8 +124,8 @@ public class AlternatingOperatorSchedule implements OperatorSchedule {
             return 0;
     }
 
-    public int getMinimumAcceptAndRejectCount() {
-        int minCount = Integer.MAX_VALUE;
+    public long getMinimumAcceptAndRejectCount() {
+        long minCount = Integer.MAX_VALUE;
         for (OperatorSchedule os : operatorSchedules) {
             if (os.getMinimumAcceptAndRejectCount() < minCount) {
                 minCount = os.getMinimumAcceptAndRejectCount();
